Some trivial type matching.

+++ src/timer.c
@@ -8,6 +8,6 @@ int get_time(struct timeval *t)
 
 uint64_t elapsed_usec(struct timeval * start_time, struct timeval * end_time)
 {
-    return ((end_time->tv_sec - start_time->tv_sec) * 1000000)
-	+ end_time->tv_usec - start_time->tv_usec;
+    return (uint64_t) ((end_time->tv_sec - start_time->tv_sec) * 1000000)
+	+ (uint64_t) end_time->tv_usec - (uint64_t) start_time->tv_usec;
 }
